backend/requirements.txt
flask
flask-sqlalchemy
openai
pytest
pytz
mockito
pytest-mockito
gunicorn
uvicorn
python-dotenv
psycopg2